Moved classes around to make them more consistent, cleaned up commented out code
