+++ yk_gmd_blender/blender/export/mesh_exporter/builders.py
@@ -1,11 +1,218 @@
 import array
 import collections
-from typing import List, Tuple, Dict, Callable, Set
+from typing import List, Tuple, Dict, Callable, Set, overload, Union, cast, Optional
+import array
+import collections
+import re
+from dataclasses import dataclass
 
 from yk_gmd_blender.yk_gmd.v2.abstract.gmd_attributes import GMDAttributeSet
 from yk_gmd_blender.yk_gmd.v2.abstract.gmd_mesh import GMDSkinnedMesh, GMDMesh
-from yk_gmd_blender.yk_gmd.v2.abstract.gmd_shader import GMDVertexBuffer, GMDVertexBufferLayout, BoneWeight4, BoneWeight
 from yk_gmd_blender.yk_gmd.v2.abstract.nodes.gmd_bone import GMDBone
+from yk_gmd_blender.yk_gmd.v2.abstract.gmd_shader import GMDVertexBuffer, GMDVertexBufferLayout, BoneWeight4, BoneWeight, VecStorage
+
+from mathutils import Vector, Matrix
+import bpy
+from bmesh.types import BMLayerItem, BMVert, BMLoop
+
+from yk_gmd_blender.yk_gmd.v2.errors.error_reporter import ErrorReporter
+
+
+@dataclass(frozen=True)
+class PerLoopVertexData:
+    normal: Optional[Vector]
+    tangent: Optional[Vector]
+    col0: Optional[Vector]
+    col1: Optional[Vector]
+    uvs: Tuple
+
+
+class VertexFetcher:
+    bm_vertices: List[BMVert]
+    vertex_layout: GMDVertexBufferLayout
+    transformation_position: Matrix
+    transformation_direction: Matrix
+    vertex_group_bone_index_map: Dict[int, int]
+    mesh: bpy.types.Mesh
+
+    # deform_layer: Optional[BMLayerItem]
+    col0_layer: Optional[bpy.types.MeshLoopColorLayer]
+    col1_layer: Optional[bpy.types.MeshLoopColorLayer]
+    tangent_layer: Optional[bpy.types.MeshLoopColorLayer]
+    tangent_w_layer: Optional[bpy.types.MeshLoopColorLayer]
+    # Stores (component length, layer)
+    uv_layers: List[Tuple[int, Optional[Union[bpy.types.MeshLoopColorLayer, bpy.types.MeshUVLoopLayer]]]]
+    error: ErrorReporter
+
+    def __init__(self,  # bm_vertices: List[BMVert],
+                 mesh_name: str,
+                 vertex_layout: GMDVertexBufferLayout,
+                 transformation_position: Matrix,
+                 transformation_direction: Matrix,
+                 mesh: bpy.types.Mesh,
+                 vertex_group_bone_index_map: Dict[int, int],
+                 # deform_layer: Optional[BMLayerItem],
+                 col0_layer: Optional[bpy.types.MeshLoopColorLayer],
+                 col1_layer: Optional[bpy.types.MeshLoopColorLayer],
+                 tangent_w_layer: Optional[bpy.types.MeshLoopColorLayer],
+                 uv_primary: Optional[bpy.types.MeshUVLoopLayer],
+                 uv_numbered: Dict[int, bpy.types.MeshLoopColorLayer],
+                 error: ErrorReporter
+                 ):
+        self.vertex_layout = vertex_layout
+        self.transformation_position = transformation_position
+        self.transformation_direction = transformation_direction
+        self.mesh = mesh
+        self.vertex_group_bone_index_map = vertex_group_bone_index_map
+
+        def verify_layer(storage, layer, name):
+            if storage:
+                if not layer:
+                    error.info(f"Expected mesh {mesh_name} to have a {name} layer but got None")
+                return layer
+            elif layer:
+                error.info(f"Mesh {mesh_name} has an unused {name} layer")
+
+        self.col0_layer = verify_layer(vertex_layout.col0_storage, col0_layer, "Color0")
+        self.col1_layer = verify_layer(vertex_layout.col1_storage, col1_layer, "Color1")
+        self.tangent_w_layer = verify_layer(
+            vertex_layout.tangent_storage in [VecStorage.Vec4Full, VecStorage.Vec4Fixed, VecStorage.Vec4Half],
+            tangent_w_layer,
+            "Tangent W Component"
+        )
+        # TODO - This "primary_uv_index" thing is icky
+        self.primary_uv_i = vertex_layout.get_primary_uv_index()
+        if self.primary_uv_i != -1 and self.primary_uv_i in uv_numbered:
+            error.recoverable(
+                f"VertexFetcher for mesh {mesh_name} given a primary uv index that refers to a numbered UV layer UV{self.primary_uv_i}. The primary UV will take precedence.")
+        self.uv_layers = []
+        for i, storage in enumerate(vertex_layout.uv_storages):
+            if self.primary_uv_i == i:
+                self.uv_layers.append((2, uv_primary))
+            elif i in uv_numbered:
+                layer = uv_numbered[i]
+                self.uv_layers.append((VecStorage.component_count(storage), layer))
+            else:
+                error.info(f"VertexFetcher for mesh {mesh_name} didn't have a UV for layer {i}, values will be all-0")
+                self.uv_layers.append((VecStorage.component_count(storage), None))
+
+        self.error = error
+
+    def normal_for(self, loop: bpy.types.MeshLoopTriangle, tri_index: int):  # , normal: Optional[Vector]
+        # normal = (self.transformation_direction @ (normal if normal is not None else Vector(loop.split_normals[tri_index]))).resized(4)
+        normal = (self.transformation_direction @ Vector(self.mesh.loops[loop.loops[tri_index]].normal)).resized(4)
+        # if self.tangent_w_layer:
+        #     # normals are stored [-1, 1] so convert from [0, 1] range
+        #     normal.w = (self.tangent_w_layer.data[loop.loops[tri_index]].color[0] * 2) - 1
+        # else:
+        normal.w = 0
+        return normal
+
+    def tangent_for(self, loop: bpy.types.MeshLoopTriangle, tri_index: int):
+        # if self.tangent_layer:
+        #     tangent = Vector(self.tangent_layer.data[loop.loops[tri_index]].color)
+        # else:
+        #     tangent = Vector((0.5, 0.5, 0.5, 0.5))
+        tangent = (self.transformation_direction @ Vector(self.mesh.loops[loop.loops[tri_index]].tangent)).resized(4)
+        if self.tangent_w_layer:
+            tangent.w = (self.tangent_w_layer.data[loop.loops[tri_index]].color[0] * 2) - 1
+        else:
+            tangent.w = 1
+        return tangent
+
+    def col0_for(self, loop: bpy.types.MeshLoopTriangle, tri_index: int):
+        if self.col0_layer:
+            return Vector(self.col0_layer.data[loop.loops[tri_index]].color)
+        else:
+            return Vector((1, 1, 1, 1))
+
+    def col1_for(self, loop: bpy.types.MeshLoopTriangle, tri_index: int):
+        if self.col1_layer:
+            return Vector(self.col1_layer.data[loop.loops[tri_index]].color)
+        else:
+            return Vector((1, 1, 1, 1))
+
+    def uv_for(self, uv_idx: int, loop: bpy.types.MeshLoopTriangle, tri_index: int):
+        component_count, layer = self.uv_layers[uv_idx]
+        if layer:
+            # If component_count == 2 then we should be storing it in a UV layer. For backwards compatibility, check if the layer actually has a "uv" section
+            if component_count == 2 and hasattr(layer.data[loop.loops[tri_index]], "uv"):
+                blender_uv = layer.data[loop.loops[tri_index]].uv
+                return Vector((blender_uv[0], 1 - blender_uv[1]))
+            else:
+                vec = Vector(layer.data[loop.loops[tri_index]].color).resized(component_count)
+                if any(x < 0 or x > 1 for x in vec):
+                    self.error.recoverable(
+                        f"UV{uv_idx} has values outside of the storable range. Expected between 0 and 1, got {vec}")
+                return Vector(layer.data[loop.loops[tri_index]].color).resized(component_count)
+        else:
+            return Vector([0] * component_count)
+
+    def boneweights_for(self, i: int):
+        # Get a list of (vertex group ID, weight) items sorted in descending order of weight
+        # Take the top 4 elements, for the top 4 most deforming bones
+        # Normalize the weights so they sum to 1
+        b_weights = sorted(list(self.mesh.vertices[i].groups), key=lambda i: i.weight, reverse=True)
+        b_weights = [(weight_pair.group, weight_pair.weight) for weight_pair in b_weights if
+                     weight_pair.group in self.vertex_group_bone_index_map]
+        if len(b_weights) > 4:
+            b_weights = b_weights[:4]
+        elif len(b_weights) < 4:
+            # Add zeroed elements to b_weights so it's 4 elements long
+            b_weights += [(0, 0.0)] * (4 - len(b_weights))
+        weight_sum = sum(weight for (vtx, weight) in b_weights)
+        if weight_sum < 0.0:
+            self.error.fatal(f"Weights {b_weights} summed to negative number!")
+        if weight_sum > 0:
+            b_weights = [(vtx_group, weight / weight_sum) for (vtx_group, weight) in b_weights]
+            # Convert the weights to the yk_gmd abstract BoneWeight format
+            weights_list = [BoneWeight(bone=self.vertex_group_bone_index_map[vtx] if weight else 0, weight=weight) for
+                            vtx, weight in
+                            b_weights]
+            return (
+                weights_list[0],
+                weights_list[1],
+                weights_list[2],
+                weights_list[3],
+            )
+        else:  # weight_sum == 0
+            return (
+                BoneWeight(bone=0, weight=0.0),
+                BoneWeight(bone=0, weight=0.0),
+                BoneWeight(bone=0, weight=0.0),
+                BoneWeight(bone=0, weight=0.0),
+            )
+
+    def get_per_loop_data(self, loop: bpy.types.MeshLoopTriangle, tri_index: int):
+        # The importer merges vertices with the same (position, boneweights, normal)
+        # The position and boneweights should match exactly the GMD, so they will always be the same for fused vertices
+        # However because Blender recalculates the normal it is also per-loop, so it might be different between fused vertices
+        return PerLoopVertexData(
+            normal=self.normal_for(loop, tri_index).freeze(),
+            tangent=self.tangent_for(loop, tri_index).freeze(),
+            col0=self.col0_for(loop, tri_index).freeze(),
+            col1=self.col1_for(loop, tri_index).freeze(),
+            uvs=tuple([self.uv_for(uv_idx, loop, tri_index).freeze() for uv_idx in range(len(self.uv_layers))])
+        )
+
+    def extract_vertex(self, vertex_buffer: GMDVertexBuffer, i: int, per_loop_data: PerLoopVertexData):
+        if vertex_buffer.layout != self.vertex_layout:
+            self.error.fatal(f"VertexFetcher told to fetch vertex for a buffer with a different layout")
+
+        vertex_buffer.pos.append((self.transformation_position @ self.mesh.vertices[i].co).resized(4))
+        if vertex_buffer.bone_weights is not None:
+            vertex_buffer.bone_weights.append(self.boneweights_for(i))
+        if vertex_buffer.normal is not None:
+            vertex_buffer.normal.append(per_loop_data.normal)
+        if vertex_buffer.tangent is not None:
+            vertex_buffer.tangent.append(per_loop_data.tangent)
+        if vertex_buffer.col0 is not None:
+            vertex_buffer.col0.append(per_loop_data.col0)
+        if vertex_buffer.col1 is not None:
+            vertex_buffer.col1.append(per_loop_data.col1)
+        for uv_idx in range(len(self.uv_layers)):
+            vertex_buffer.uvs[uv_idx].append(per_loop_data.uvs[uv_idx])
+
 
 class SubmeshBuilder:
     """
@@ -37,22 +244,6 @@ class SubmeshBuilder:
             return idx
 
         return self.blender_vid_to_this_vid[overall_blender_idx]
-
-
-    # Adds the vertex if not already present
-    # Used for smooth faces, where the vertex is reused in other faces as well
-    # def add_vertex(self, blender_vid, generate_vertex: Callable[[GMDVertexBuffer], None]) -> int:
-    #     if blender_vid not in self.blender_vid_to_this_vid:
-    #         self.blender_vid_to_this_vid[blender_vid] = self.add_unique_vertex(generate_vertex)
-    #
-    #     return self.blender_vid_to_this_vid[blender_vid]
-    #
-    # # Adds a unique vertex, which we assume is never duplicated
-    # # Used for hard edges
-    # def add_unique_vertex(self, generate_vertex: Callable[[GMDVertexBuffer], None]) -> int:
-    #     idx = len(self.vertices)
-    #     generate_vertex(self.vertices)
-    #     return idx
 
     # Calls the function to add the vertex to the buffer and returns the index of the generated vertex.
     # This is anonymous, it will not be considered for per-loop duplication and will not be added to any vertex ID counts
